+++ src/main/java/handalab/eggtec/EggtecApplication.java
@@ -7,7 +7,6 @@ import org.springframework.scheduling.annotation.EnableScheduling;
 
 @SpringBootApplication
 @EnableScheduling
-@EnableAsync
 public class EggtecApplication {
 
 	public static void main(String[] args) {
