--- src/main/java/handalab/eggtec/dto/JsonDTO.java
@@ -1,10 +1,0 @@
-package handalab.eggtec.dto;
-
-import lombok.Data;
-
-@Data
-public class JsonDTO {
-    private String type;
-    private String value;
-    private boolean isNull;
-}
