--- src/main/java/handalab/eggtec/dto/request/history/CsvFilterDTO.java
@@ -1,15 +1,0 @@
-package handalab.eggtec.dto.request.history;
-
-import com.fasterxml.jackson.annotation.JsonProperty;
-import lombok.Data;
-
-import java.time.LocalDate;
-
-@Data
-public class CsvFilterDTO {
-    @JsonProperty("start_date")
-    private LocalDate startDate;
-    @JsonProperty("end_date")
-    private LocalDate endDate;
-    private String path;
-}
