+++ src/main/java/handalab/eggtec/dto/setting/SelectedDataDTO.java
@@ -1,22 +1,19 @@
-package handalab.eggtec.dto.response.setting;
+package handalab.eggtec.dto.setting;
 
 import com.fasterxml.jackson.annotation.JsonInclude;
 import com.fasterxml.jackson.annotation.JsonProperty;
 import lombok.Data;
 
+import java.util.List;
+
 @Data
 @JsonInclude(JsonInclude.Include.NON_NULL)
-public class SettingDTO {
-    private Integer idx;
-    @JsonProperty("output_cnt")
+public class SelectedDataDTO {
     private Integer outputCnt;
-    private String title;
-    @JsonProperty("exp_white")
-    private Integer expWhite;
-    @JsonProperty("exp_brown")
-    private Integer expBrown;
+    private Integer exposure;
     private Integer y1;
     private Integer y2;
     private Integer y3;
     private Integer y4;
+
 }
