+++ src/main/java/handalab/eggtec/mapper/HistoryMapper.java
@@ -1,9 +1,6 @@
 package handalab.eggtec.mapper;
 
-import handalab.eggtec.dto.request.history.CsvFilterDTO;
-import handalab.eggtec.dto.request.history.HistoryFilterDTO;
-import handalab.eggtec.dto.response.history.HistoryDTO;
-import handalab.eggtec.dto.response.history.*;
+import handalab.eggtec.dto.history.*;
 import org.apache.ibatis.annotations.Mapper;
 import org.apache.ibatis.annotations.Param;
 
@@ -11,18 +8,15 @@ import java.util.List;
 
 @Mapper
 public interface HistoryMapper {
-//    @Select("SELECT * FROM history WHERE id = #{id}")
-    List<HistoryTestDTO> getHistory(int id);
-
     List<TotalSummaryDTO> getTotalSummary();
 
-    List<SummaryDTO> getSummary(@Param("id") Integer id, @Param("filter") HistoryFilterDTO filter); //Integer?
+    List<SummaryDTO> getSummary(@Param("id") Integer id, @Param("filter") CsvDTO.HistoryFilterDTO filter); //Integer?
 
     List<LastDTO> getLast();
 
     List<TotalDTO> getTotal();
 
-    List<CsvDTO> getCsvData(@Param("id") Integer id, @Param("filter") CsvFilterDTO info);
+    List<CsvDTO> getCsvData(@Param("id") Integer id, @Param("filter") CsvDTO.CsvFilterDTO info);
 
     HistoryDTO postHistory(@Param("history") HistoryDTO history);
