+++ src/main/java/handalab/eggtec/socketio/IntervalService.java
@@ -1,18 +1,13 @@
 package handalab.eggtec.socketio;
 
 import com.corundumstudio.socketio.SocketIOServer;
-import handalab.eggtec.dto.MessageDTO;
-import handalab.eggtec.dto.response.setting.SettingDTO;
 import handalab.eggtec.dto.socketio.IntervalDTO;
-import handalab.eggtec.dto.socketio.SocketMsgDTO;
 import handalab.eggtec.mapper.SocketMapper;
-import lombok.Getter;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.scheduling.concurrent.ThreadPoolTaskScheduler;
 import org.springframework.stereotype.Service;
 
 import java.time.Duration;
-import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 import java.util.concurrent.ScheduledFuture;
@@ -22,7 +17,6 @@ public class IntervalService {
 
     @Value("${socketio.server.port.interval}")
     private Integer intervalPort;
-//    private Integer intervalPort = 5004;
 
     private final Map<Integer, SocketIOServer> socketIOServers;
     private final SocketMapper socketMapper;
@@ -64,70 +58,3 @@ public class IntervalService {
         return isRunning;
     }
 }
-
-/*   @Scheduled
-package handalab.eggtec.socketio;
-
-import com.corundumstudio.socketio.SocketIOServer;
-import handalab.eggtec.dto.MessageDTO;
-import handalab.eggtec.dto.response.setting.SettingDTO;
-import handalab.eggtec.dto.socketio.IntervalDTO;
-import handalab.eggtec.dto.socketio.SocketMsgDTO;
-import handalab.eggtec.mapper.SocketMapper;
-import jakarta.annotation.PreDestroy;
-import lombok.Getter;
-import org.springframework.beans.factory.annotation.Value;
-import org.springframework.scheduling.annotation.Async;
-import org.springframework.scheduling.annotation.Scheduled;
-import org.springframework.scheduling.concurrent.ThreadPoolTaskScheduler;
-import org.springframework.stereotype.Service;
-
-import java.time.Duration;
-import java.util.HashMap;
-import java.util.List;
-import java.util.Map;
-import java.util.concurrent.ScheduledFuture;
-
-@Service
-public class IntervalService {
-
-//    @Value("${socketio.server.port.interval}")
-//    private Integer intervalPort;
-    private Integer intervalPort = 5004;
-
-    private final Map<Integer, SocketIOServer> socketIOServers;
-    private final SocketMapper socketMapper;
-
-//    private final ThreadPoolTaskScheduler taskScheduler;
-//    private ScheduledFuture<?> scheduledTask;
-    private boolean isRunning = false;
-
-    public IntervalService(SocketMapper socketMapper, Map<Integer, SocketIOServer> socketIOServer) {
-//        this.taskScheduler = new ThreadPoolTaskScheduler();
-//        this.taskScheduler.initialize();
-        this.socketMapper = socketMapper;
-        this.socketIOServers = socketIOServer;
-    }
-
-    public void startInterval() {
-        isRunning = true;
-    }
-
-    public void stopInterval() {
-        isRunning = false;
-    }
-
-    @Async
-    @Scheduled(fixedRate = 2000)
-    void executeTask() {
-        if(!isRunning) {return;}
-        List<IntervalDTO> intervalDTO = socketMapper.getIntervalData();
-        socketIOServers.get(intervalPort).getBroadcastOperations().sendEvent("message", intervalDTO);
-    }
-
-    public boolean isRunning() {
-        return isRunning;
-    }
-
-}
- */
