+++ src/main/resources/mappers/recipeMapper.xml
@@ -7,22 +7,22 @@
         >
 
 <mapper namespace="handalab.eggtec.mapper.RecipeMapper">
-    <select id = "getAllRecipes" resultType="handalab.eggtec.dto.response.recipe.RecipeDTO">
+    <select id = "getAllRecipes" resultType="handalab.eggtec.dto.recipe.RecipeDTO">
         SELECT * FROM recipe ORDER BY idx
     </select>
 
 
-    <select id = "getRecipe" resultType="handalab.eggtec.dto.response.recipe.RecipeDTO">
+    <select id = "getRecipe" resultType="handalab.eggtec.dto.recipe.RecipeDTO">
         SELECT * FROM recipe where idx = #{id};
     </select>
 
-    <select id = "postRecipe" resultType="handalab.eggtec.dto.response.recipe.RecipeDTO">
+    <select id = "postRecipe" resultType="handalab.eggtec.dto.recipe.RecipeDTO">
         INSERT INTO recipe (idx, nickname, type, weight)
         VALUES (#{recipe.idx}, #{recipe.nickname}, #{recipe.type}, #{recipe.weight})
             RETURNING *;
     </select>
 
-    <select id = "updateRecipe" resultType = "handalab.eggtec.dto.response.recipe.RecipeDTO">
+    <select id = "updateRecipe" resultType = "handalab.eggtec.dto.recipe.RecipeDTO">
         UPDATE recipe
         <set>
             <if test = "r.idx != null">idx = #{r.idx},</if>
@@ -35,7 +35,7 @@
         RETURNING *;
     </select>
 
-    <select id = "deleteRecipe" resultType="handalab.eggtec.dto.response.recipe.RecipeDTO">
+    <select id = "deleteRecipe" resultType="handalab.eggtec.dto.recipe.RecipeDTO">
         DELETE FROM recipe WHERE idx = #{id} RETURNING *
     </select>
